+++ app/src/main/java/app/revanced/integrations/shared/settings/preference/SharedPrefCategory.java
@@ -7,8 +7,8 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import app.revanced.integrations.shared.Logger;
 import app.revanced.integrations.shared.Utils;
-
-import java.util.Objects;
+import org.json.JSONObject;
+import java.util.*;
 
 /**
  * Shared categories, and helper methods.
@@ -43,6 +43,10 @@ public class SharedPrefCategory {
         preferences.edit().putBoolean(key, value).apply();
     }
 
+    public void saveSet(@NonNull String key, Set<String> value) {
+        preferences.edit().putStringSet(key, value).apply();
+    }
+
     /**
      * @param value a NULL parameter removes the value from the preferences
      */
@@ -67,6 +71,16 @@ public class SharedPrefCategory {
     /**
      * @param value a NULL parameter removes the value from the preferences
      */
+
+    public Set<String> getSet(@NonNull String key, Set<String> _default) {
+        try {
+            return preferences.getStringSet(key, _default);
+        } catch (ClassCastException ex) {
+            // Value stored is a completely different type (should never happen).
+            removeConflictingPreferenceKeyValue(key);
+            return _default;
+        }
+    }
     public void saveString(@NonNull String key, @Nullable String value) {
         saveObjectAsString(key, value);
     }
@@ -146,6 +160,15 @@ public class SharedPrefCategory {
                 removeConflictingPreferenceKeyValue(key);
                 return _default;
             }
+        }
+    }
+
+    public JSONObject getAll(){
+        try{
+            Map<String, ?> allEntries = preferences.getAll();
+            return new JSONObject(allEntries);
+        }catch (Exception ex){
+            return new JSONObject();
         }
     }