+++ CHANGELOG.md
@@ -1,3 +1,94 @@
+## [1.14.0-dev.1](https://github.com/crimera/revanced-integrations/compare/v1.13.0...v1.14.0-dev.1) (2024-04-19)
+
+
+### Features
+
+* **Twitter:** Added `Profile tabs customisation` ([2a498d3](https://github.com/crimera/revanced-integrations/commit/2a498d365390d467d54d6073fba1bd2a71e39f91))
+
+## [1.13.0](https://github.com/crimera/revanced-integrations/compare/v1.12.0...v1.13.0) (2024-04-19)
+
+
+### Bug Fixes
+
+* Make the methods `notifyItemChanged` and `getAdapterPosition` search for the functions dynamically ([ff144f7](https://github.com/crimera/revanced-integrations/commit/ff144f79ce2e9645f8bd680f12a3329c2f74e3e2))
+* Use `piko_pref_feature_flags` for the Feature Flags button ([2afce73](https://github.com/crimera/revanced-integrations/commit/2afce734d7ce7d0d713beda1b9e82d6a11818698))
+
+
+### Features
+
+* **Twitter:** Added ability to export and import prefs & feature flags ([f87d4ed](https://github.com/crimera/revanced-integrations/commit/f87d4edfedeb77cd459e02f3513b498fc9b461db))
+* **Twitter:** Added ability to export and import prefs & feature flags ([e13fc7e](https://github.com/crimera/revanced-integrations/commit/e13fc7ef7a7c99634b3b7ccbcf836428556fccc9))
+
+
+### Updates
+
+* Use `piko_pref_feature_flags` in toolbar ([9e11984](https://github.com/crimera/revanced-integrations/commit/9e11984563b8edea410ecfbecb1e6f6646cb07e7))
+
+
+### Refactors
+
+* **Twitter:** load `Utils.load()` dynamically ([6cddfa7](https://github.com/crimera/revanced-integrations/commit/6cddfa714088a081b6de75fcfbb8a8cee1445f22))
+
+## [1.13.0-dev.4](https://github.com/crimera/revanced-integrations/compare/v1.13.0-dev.3...v1.13.0-dev.4) (2024-04-18)
+
+
+### Refactors
+
+* **Twitter:** load `Utils.load()` dynamically ([6cddfa7](https://github.com/crimera/revanced-integrations/commit/6cddfa714088a081b6de75fcfbb8a8cee1445f22))
+
+## [1.13.0-dev.3](https://github.com/crimera/revanced-integrations/compare/v1.13.0-dev.2...v1.13.0-dev.3) (2024-04-18)
+
+
+### Bug Fixes
+
+* Make the methods `notifyItemChanged` and `getAdapterPosition` search for the functions dynamically ([ff144f7](https://github.com/crimera/revanced-integrations/commit/ff144f79ce2e9645f8bd680f12a3329c2f74e3e2))
+
+## [1.13.0-dev.2](https://github.com/crimera/revanced-integrations/compare/v1.13.0-dev.1...v1.13.0-dev.2) (2024-04-17)
+
+
+### Updates
+
+* Use `piko_pref_feature_flags` in toolbar ([9e11984](https://github.com/crimera/revanced-integrations/commit/9e11984563b8edea410ecfbecb1e6f6646cb07e7))
+
+## [1.13.0-dev.1](https://github.com/crimera/revanced-integrations/compare/v1.12.1-dev.1...v1.13.0-dev.1) (2024-04-15)
+
+
+### Features
+
+* **Twitter:** Added ability to export and import prefs & feature flags ([f87d4ed](https://github.com/crimera/revanced-integrations/commit/f87d4edfedeb77cd459e02f3513b498fc9b461db))
+* **Twitter:** Added ability to export and import prefs & feature flags ([e13fc7e](https://github.com/crimera/revanced-integrations/commit/e13fc7ef7a7c99634b3b7ccbcf836428556fccc9))
+
+## [1.12.1-dev.1](https://github.com/crimera/revanced-integrations/compare/v1.12.0...v1.12.1-dev.1) (2024-04-15)
+
+
+### Bug Fixes
+
+* Use `piko_pref_feature_flags` for the Feature Flags button ([2afce73](https://github.com/crimera/revanced-integrations/commit/2afce734d7ce7d0d713beda1b9e82d6a11818698))
+
+## [1.12.0](https://github.com/crimera/revanced-integrations/compare/v1.11.0...v1.12.0) (2024-04-14)
+
+
+### Features
+
+* Add support for adding custom feature flags [#28](https://github.com/crimera/revanced-integrations/issues/28) ([fe9ca52](https://github.com/crimera/revanced-integrations/commit/fe9ca527cbc1c93d7df674d69876fe0d22219e42))
+* Added resource file for strings ([cc3583d](https://github.com/crimera/revanced-integrations/commit/cc3583d88396d1e0f0f63ce6a42a53dfd9fc8df7))
+* Switched preferences to custom file ([3f197da](https://github.com/crimera/revanced-integrations/commit/3f197da08a13fe5a542ba3b2ee321774726f0a1c))
+* **Twitter:** Added `Feature Flag Hook` ([fd1c439](https://github.com/crimera/revanced-integrations/commit/fd1c4390ce2a63f13f38a81c7d66471a510ee71f))
+* **Twitter:** Added `Hide immersive player` patch ([86e09a8](https://github.com/crimera/revanced-integrations/commit/86e09a8dc0f24040ddd05e78eecc5bb2a21f7675))
+
+
+### Updates
+
+* **Hook feature flag:** Save flags on modify list ([3537091](https://github.com/crimera/revanced-integrations/commit/3537091036b29a9eed27fcad8dbe00d36b8de6aa))
+* **Twitter:** Made some features to use flags ([e4bb878](https://github.com/crimera/revanced-integrations/commit/e4bb87834258f95fbb1a58a98d9292865c4a2748))
+
+## [1.12.0-dev.5](https://github.com/crimera/revanced-integrations/compare/v1.12.0-dev.4...v1.12.0-dev.5) (2024-04-14)
+
+
+### Updates
+
+* **Hook feature flag:** Save flags on modify list ([3537091](https://github.com/crimera/revanced-integrations/commit/3537091036b29a9eed27fcad8dbe00d36b8de6aa))
+
 ## [1.12.0-dev.4](https://github.com/crimera/revanced-integrations/compare/v1.12.0-dev.3...v1.12.0-dev.4) (2024-04-14)